eval/semantic_chunking/requirements.txt
langchain == 0.3.0
langchain-openai == 0.2.0
langchain-experimental == 0.3.0
python-dotenv == 1.0.1